Allow the site to be built from Git
Before, the MP3 files were not included in the repository (because they
did not fit on GitHub). Instead, when changing hosting providers, I
simply packed up all the docroot files as they were and copied them over
to the new hoster, without running `mksite.sh` anymore at all.

However, I don't want to be dependent on backups or manually moving
these files around anymore. You should be able to `git clone`, run
`mksite.sh` and be done. And that's exactly how it's working now.

I know that GitHub isn't really the best place to host a podcast, and
yes, the repo is 500M in size now. On the other hand, I don't want to
pay $5/mo just for hosting a podcast that's 10 years old now and only
lasted for two months, so I hope GitHub will forgive me for not using
Git LFS for it.